CustomLink rewritten to TS
