+++ src/components/CustomLink/CustomLink.mdx
@@ -1,5 +1,5 @@
 import { Playground, Props } from 'docz'
-import CustomLink from '@components/CustomLink'
+import { CustomLink } from '@components/CustomLink'
 
 # CustomLink
