+++ src/components/CustomLink/CustomLink.test.js
@@ -1,8 +1,7 @@
 import '@testing-library/jest-dom'
 
-import CustomLink from '@components/CustomLink'
+import { CustomLink } from '@components/CustomLink'
 import { fireEvent, render } from '@testing-library/react'
-import React from 'react'
 
 describe('<CustomLink/> tests', () => {
   const props = { label: 'Custom link' }
