+++ src/components/CustomLink/index.ts
@@ -1,0 +1,1 @@
+export * from './CustomLink'
