+++ src/components/index.ts
@@ -15,7 +15,7 @@ export { default as CheckboxNoFormik } from './Checkbox'
 export * from './Close'
 export * from './ColorLine'
 export { default as ColorNumber } from './ColorNumber'
-export { default as CustomLink } from './CustomLink'
+export * from './CustomLink'
 export { default as DateTimePicker } from './DateTimePicker'
 export { default as Divider } from './Divider'
 export { default as Dot } from './Dot'
